+++ tests/pretrained_test.py
@@ -4,15 +4,18 @@ import pytest
 import spacy
 
 from allennlp.common.testing import AllenNlpTestCase
-
+from allennlp.predictors import Predictor
 from allennlp_models import pretrained
 
 
 # But default we don't run these tests
 @pytest.mark.pretrained_model_test
 class TestAllenNlpPretrained(AllenNlpTestCase):
+    def setup(self):
+        self.pretrained_models = pretrained.get_pretrained_models()
+
     def test_machine_comprehension(self):
-        predictor = pretrained.bidirectional_attention_flow_seo_2017()
+        predictor = Predictor.from_path(self.pretrained_models["rc-bidaf"].archive_file)
 
         passage = """The Matrix is a 1999 science fiction action film written and directed by The Wachowskis, starring Keanu Reeves, Laurence Fishburne, Carrie-Anne Moss, Hugo Weaving, and Joe Pantoliano. It depicts a dystopian future in which reality as perceived by most humans is actually a simulated reality called "the Matrix", created by sentient machines to subdue the human population, while their bodies' heat and electrical activity are used as an energy source. Computer programmer Neo" learns this truth and is drawn into a rebellion against the machines, which involves other people who have been freed from the "dream world". """
         question = "Who stars in The Matrix?"
@@ -26,7 +29,9 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         assert correct == result["best_span_str"]
 
     def test_semantic_role_labeling(self):
-        predictor = pretrained.bert_srl_shi_2019()
+        predictor = Predictor.from_path(
+            self.pretrained_models["structured-prediction-srl-bert"].archive_file
+        )
 
         sentence = "If you liked the music we were playing last night, you will absolutely love what we're playing tomorrow!"
 
@@ -199,7 +204,9 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         ]
 
     def test_textual_entailment(self):
-        predictor = pretrained.decomposable_attention_with_elmo_parikh_2017()
+        predictor = Predictor.from_path(
+            self.pretrained_models["pair-classification-decomposable-attention-elmo"].archive_file
+        )
 
         result = predictor.predict_json(
             {
@@ -229,7 +236,7 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         assert result["label_probs"][2] > 0.6  # neutral
 
     def test_coreference_resolution(self):
-        predictor = pretrained.neural_coreference_resolution()
+        predictor = Predictor.from_path(self.pretrained_models["coref-spanbert"].archive_file)
 
         document = "We 're not going to skimp on quality , but we are very focused to make next year . The only problem is that some of the fabrics are wearing out - since I was a newbie I skimped on some of the fabric and the poor quality ones are developing holes ."
 
@@ -297,7 +304,9 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         ]
 
     def test_ner(self):
-        predictor = pretrained.named_entity_recognition_with_elmo_peters_2018()
+        predictor = Predictor.from_path(
+            self.pretrained_models["tagging-elmo-crf-tagger"].archive_file
+        )
 
         sentence = """Michael Jordan is a professor at Berkeley."""
 
@@ -320,7 +329,9 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         reason="this model changed before and after 2.1 and 2.2",
     )
     def test_constituency_parsing(self):
-        predictor = pretrained.span_based_constituency_parsing_with_elmo_joshi_2018()
+        predictor = Predictor.from_path(
+            self.pretrained_models["structured-prediction-constituency-parser"].archive_file
+        )
 
         sentence = """Pierre Vinken died aged 81; immortalised aged 61."""
 
@@ -344,7 +355,9 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         )
 
     def test_dependency_parsing(self):
-        predictor = pretrained.biaffine_parser_universal_dependencies_todzat_2017()
+        predictor = Predictor.from_path(
+            self.pretrained_models["structured-prediction-biaffine-parser"].archive_file
+        )
         sentence = """He ate spaghetti with chopsticks."""
         result = predictor.predict_json({"sentence": sentence})
         # Note that this tree is incorrect. We are checking here that the decoded
@@ -365,7 +378,9 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         assert result["predicted_heads"] == [2, 0, 2, 2, 4, 2]
 
     def test_openie(self):
-        predictor = pretrained.open_information_extraction_stanovsky_2018()
+        predictor = Predictor.from_path(
+            self.pretrained_models["structured-prediction-srl"].archive_file
+        )
         result = predictor.predict_json(
             {"sentence": "I'm against picketing, but I don't know how to show it."}
         )
@@ -373,15 +388,17 @@ class TestAllenNlpPretrained(AllenNlpTestCase):
         assert "words" in result
 
     @pytest.mark.parametrize(
-        "get_model_fn",
-        [
-            pretrained.fine_grained_named_entity_recognition,
-            pretrained.fine_grained_named_entity_recognition_transformer,
-        ],
+        "get_model_arg",
+        ["tagging-fine-grained-crf-tagger", "tagging-fine-grained-transformer-crf-tagger"],
     )
-    def test_fine_grained_ner(self, get_model_fn):
-        predictor = get_model_fn()
+    def test_fine_grained_ner(self, get_model_arg):
+        predictor = Predictor.from_path(self.pretrained_models[get_model_arg].archive_file)
         text = """Dwayne Haskins passed for 251 yards and three touchdowns, and Urban Meyer finished his coaching career at Ohio State with a 28-23 victory after the Buckeyes held off Washington’s thrilling fourth-quarter comeback in the 105th Rose Bowl on Tuesday. Parris Campbell, Johnnie Dixon and Rashod Berry caught TD passes in the first half for the fifth-ranked Buckeyes (13-1), who took a 25-point lead into the fourth. But Myles Gaskin threw a touchdown pass and rushed for two more scores for the No. 9 Huskies (10-4), scoring from 2 yards out with 42 seconds left. The Buckeyes intercepted Jake Browning’s pass on the 2-point conversion attempt and then recovered the Huskies’ onside kick to wrap up the final game of Meyer’s seven-year tenure. “I’m a very blessed man,” Meyer said. “I’m blessed because of my family, [but] this team, this year, I love this group as much as any I’ve ever had.”"""
         result = predictor.predict_json({"sentence": text})
         # Just assert that we predicted something better than all-O.
         assert len(frozenset(result["tags"])) > 1
+
+    def test_pretrained_models_archives(self):
+        for key, val in self.pretrained_models.items():
+            # Each model in pretrained_models should have an archive.
+            assert val.archive_file